repository implementation with jetbrain exposed

+++ src/main/kotlin/no/itverket/batchdemo/dal/Person.kt
@@ -8,6 +8,6 @@ data class Person(
     val company: String,
     val street: String,
     val city: String,
-    val zipCode: Int,
+    val zipCode: Short,
     val phone: String
 )
